requirements.txt
streamlit==1.29.0
supabase==1.0.3
google-generativeai==0.8.0
python-dotenv==1.0.0
pandas==2.1.4
